Implementing validation using Joi

+++ src/controllers/RoomController.ts
@@ -1,6 +1,6 @@
 //  Importing express, mongoose and JWT resources
 import { Request, Response } from "express";
-import mongoose from "mongoose";
+import { isValidObjectId } from "mongoose";
 
 //	Importing Rooms repository
 import RoomsRepository from "../repositories/RoomsRepository";
@@ -16,7 +16,7 @@ class RoomController {
 	async index(req: Request, res: Response) {
 		const userId = req.body.user.id;
 
-		if(!userId || !userId.length || !mongoose.isValidObjectId(userId)) {
+		if(!userId || !userId.length || !isValidObjectId(userId)) {
 			return res.status(400).send("Invalid id!");
 		}
 
@@ -35,6 +35,10 @@ class RoomController {
 	async create(req: Request, res: Response) {
 		const userId = req.body.user.id;
 		const { name } = req.body;
+
+		if(!userId || !userId.length || !isValidObjectId(userId)) {
+			return res.status(400).send("Invalid id!");
+		}
 
 		await RoomsRepository.create({
 			userId,
@@ -68,6 +72,14 @@ class RoomController {
 		const roomId = req.params.id;
 		const { name } = req.body;
 
+		if(!userId || !userId.length || !isValidObjectId(userId)) {
+			return res.status(400).send("Invalid id!");
+		}
+
+		if(!roomId || !roomId.length || !isValidObjectId(roomId)) {
+			return res.status(400).send("Invalid id!");
+		}
+
 		await RoomsRepository.update(roomId, userId, {
 			name: name.trim()
 		}).then((response) => {
@@ -86,6 +98,24 @@ class RoomController {
 		const userId = req.body.user.id;
 		const roomId = req.params.id;
 		const filename = (req.file) ? req.file.filename : "";
+
+		if(filename) {
+			const mimeType = (req?.file?.mimetype) ? req?.file?.mimetype.split("/")[0] : null;
+
+			if(!mimeType || !mimeType.length || (mimeType !== "image")) {
+				return res.status(400).send("Invalid image type!");
+			}
+		} else {
+			return res.status(400).send("Invalid image!");
+		}
+
+		if(!userId || !userId.length || !isValidObjectId(userId)) {
+			return res.status(400).send("Invalid id!");
+		}
+
+		if(!roomId || !roomId.length || !isValidObjectId(roomId)) {
+			return res.status(400).send("Invalid id!");
+		}
 
 		await RoomsRepository.findByIds(roomId, userId).then((room) => {
 			if(room) {
@@ -126,6 +156,14 @@ class RoomController {
 		const userId = req.body.user.id;
 		const roomId = req.params.id;
 
+		if(!userId || !userId.length || !isValidObjectId(userId)) {
+			return res.status(400).send("Invalid id!");
+		}
+
+		if(!roomId || !roomId.length || !isValidObjectId(roomId)) {
+			return res.status(400).send("Invalid id!");
+		}
+
 		await RoomsRepository.delete(roomId, userId).then((room) => {
 			if(room) {
 				if(room.image && room.image.length) {
@@ -145,6 +183,14 @@ class RoomController {
 	async search(req: Request, res: Response) {
 		const userId = req.body.user.id;
 		const query = req.query.q;
+
+		if(!userId || !userId.length || !isValidObjectId(userId)) {
+			return res.status(400).send("Invalid id!");
+		}
+
+		if(!query || !query.length) {
+			return res.status(400).send("Invalid query!");
+		}
 
 		await RoomsRepository.find(userId, <string>query).then((response) => {
 			if(response) {
